Crate network_node

Common utilities and traits for the dr_ones network simulator.

This crate provides shared functionality used by the drone, client, and server components of the network simulator.

Re-exports

pub use logging::disable_logging;

pub use logging::enable_logging;

pub use logging::is_logging_enabled;

pub use logging::redirect_logs_to_file;

Modules

logging

Macros

log_error

Logs an error message.

log_status

Logs a status message.

Enums

ClientCommand

Command

SerializableMessage

ServerCommand

ServerType

Traits

NetworkNode

Common network functionality shared across different node types. This trait provides basic network operations that all network nodes (drones, clients, and servers) need to implement.